+++ src/utils/createPasswordValidator.ts
@@ -1,21 +1,31 @@
-// src/utils/createPasswordValidator.ts
+// nist-password-validator\src\utils\createPasswordValidator.ts
+// src/utils/PasswordValidator.ts
 import { validatePassword } from "../validatePassword ";
 import { ValidationOptions, ValidationResult } from "../types";
 
-/**
- * Creates a reusable password validator with pre-configured options.
- * @param {ValidationOptions} options - The validation options to be used for all validations with this validator.
- * @returns {object} - An object with a `validate` method to validate passwords using the configured options.
- */
-export function createPasswordValidator(options: ValidationOptions = {}) {
-  return {
-    /**
-     * Validates a password using the pre-configured options.
-     * @param {string} password - The password to validate.
-     * @returns {Promise<ValidationResult>} - A promise that resolves to the validation result.
-     */
-    validate: (password: string): Promise<ValidationResult> => {
-      return validatePassword(password, options);
-    },
-  };
+class PasswordValidator {
+  private options: ValidationOptions;
+
+  constructor(options: ValidationOptions = {}) {
+    this.options = options;
+  }
+
+  /**
+   * Updates the configuration options.
+   * @param {ValidationOptions} options - New validation options.
+   */
+  updateConfig(options: ValidationOptions): void {
+    this.options = { ...this.options, ...options };
+  }
+
+  /**
+   * Validates a password using the configured options.
+   * @param {string} password - The password to validate.
+   * @returns {Promise<ValidationResult>} - A promise that resolves to the validation result.
+   */
+  async validate(password: string): Promise<ValidationResult> {
+    return validatePassword(password, this.options);
+  }
 }
+
+export { PasswordValidator };